+++ tests/test_interpolation.py
@@ -5,10 +5,10 @@ import numpy as np
 from plonk.visualize.interpolation import scalar_interpolation, vector_interpolation
 
 from .stubdata.interpolation_arrays import (
-    scalar_cross_section,
     scalar_projection,
-    vector_cross_section,
+    scalar_slice,
     vector_projection,
+    vector_slice,
 )
 
 N = 10
@@ -46,7 +46,7 @@ def test_scalar_interpolation_projection():
     np.testing.assert_allclose(im, scalar_projection, rtol=1e-5)
 
 
-def test_scalar_interpolation_cross_section():
+def test_scalar_interpolation_slice():
     """Test cross section interpolation."""
     im = scalar_interpolation(
         quantity=S_DATA,
@@ -60,7 +60,7 @@ def test_scalar_interpolation_cross_section():
         number_of_pixels=PIX,
     )
 
-    np.testing.assert_allclose(im, scalar_cross_section, rtol=1e-5)
+    np.testing.assert_allclose(im, scalar_slice, rtol=1e-5)
 
 
 def test_vector_interpolation_projection():
@@ -80,7 +80,7 @@ def test_vector_interpolation_projection():
     np.testing.assert_allclose(vec, vector_projection, rtol=1e-5)
 
 
-def test_vector_interpolation_cross_section():
+def test_vector_interpolation_slice():
     """Test cross section interpolation."""
     vec = vector_interpolation(
         quantity_x=X_DATA,
@@ -95,4 +95,4 @@ def test_vector_interpolation_cross_section():
         number_of_pixels=PIX,
     )
 
-    np.testing.assert_allclose(vec, vector_cross_section, rtol=1e-5)
+    np.testing.assert_allclose(vec, vector_slice, rtol=1e-5)